requirements.txt
blinker==1.8.2
click==8.1.8
flask==3.0.3
gunicorn==23.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
packaging==24.2
python-dotenv==1.0.1
werkzeug==3.0.6
zipp==3.20.2
Faker==37.1.0
iniconfig==2.0.0
pluggy==1.5.0
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
six==1.17.0
tomli==2.2.1
typing-extensions==4.12.2
werkzeug==3.0.6
zipp==3.20.2
flask-login==0.6.3
mysql-connector-python==8.1.0
SQLAlchemy==2.0.30

Flask-Migrate
